+++ mod-pubsub-server/src/main/java/org/folio/config/ApplicationConfig.java
@@ -1,12 +1,36 @@
 package org.folio.config;
 
+import io.vertx.core.Vertx;
+import io.vertx.kafka.client.producer.KafkaProducer;
+import org.apache.kafka.clients.admin.AdminClient;
+import org.apache.kafka.clients.admin.AdminClientConfig;
+import org.folio.kafka.KafkaConfig;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+
+import java.util.HashMap;
+import java.util.Map;
 
 @Configuration
 @ComponentScan(basePackages = {
   "org.folio.dao",
   "org.folio.services",
-  "org.folio.rest"})
+  "org.folio.rest",
+  "org.folio.kafka"})
 public class ApplicationConfig {
+
+  @Bean
+  public KafkaProducer kafkaProducer(@Autowired Vertx vertx, @Autowired KafkaConfig config) {
+    return KafkaProducer.createShared(vertx, "pub-sub-producer", config.getProducerProps());
+  }
+
+  @Bean
+  public AdminClient adminClient(@Autowired KafkaConfig config) {
+    Map<String, Object> configs = new HashMap<>();
+    configs.put(AdminClientConfig.BOOTSTRAP_SERVERS_CONFIG, config.getUrl());
+    return AdminClient.create(configs);
+  }
+
 }