+++ mod-pubsub-server/src/main/java/org/folio/kafka/KafkaConfig.java
@@ -1,0 +1,49 @@
+package org.folio.kafka;
+
+import org.apache.kafka.clients.consumer.ConsumerConfig;
+import org.apache.kafka.clients.producer.ProducerConfig;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Component;
+
+import java.util.HashMap;
+import java.util.Map;
+
+@Component
+public class KafkaConfig {
+
+  @Value("${KAFKA_HOST}")
+  private String kafkaHost;
+  @Value("${KAFKA_PORT}")
+  private String kafkaPort;
+
+  public String getKafkaHost() {
+    return kafkaHost;
+  }
+
+  public String getKafkaPort() {
+    return kafkaPort;
+  }
+
+  public Map<String, String> getProducerProps() {
+    Map<String, String> producerProps = new HashMap<>();
+    producerProps.put(ProducerConfig.BOOTSTRAP_SERVERS_CONFIG, getUrl());
+    producerProps.put(ProducerConfig.ENABLE_IDEMPOTENCE_CONFIG, "true");
+    producerProps.put(ProducerConfig.KEY_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
+    producerProps.put(ProducerConfig.VALUE_SERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringSerializer");
+    return producerProps;
+  }
+
+  public Map<String, String> getConsumerProps() {
+    Map<String, String> consumerProps = new HashMap<>();
+    consumerProps.put(ConsumerConfig.BOOTSTRAP_SERVERS_CONFIG, getUrl());
+    consumerProps.put(ConsumerConfig.ENABLE_AUTO_COMMIT_CONFIG, "true");
+    consumerProps.put(ConsumerConfig.AUTO_COMMIT_INTERVAL_MS_CONFIG, "1000");
+    consumerProps.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer");
+    consumerProps.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, "org.apache.kafka.common.serialization.StringDeserializer");
+    return consumerProps;
+  }
+
+  public String getUrl() {
+    return kafkaHost + ":" + kafkaPort;
+  }
+}