+++ mod-pubsub-server/src/test/java/org/folio/config/TestConfig.java
@@ -1,0 +1,7 @@
+package org.folio.config;
+
+import org.springframework.context.annotation.Import;
+
+@Import(ApplicationConfig.class)
+public class TestConfig {
+}